# search: Modify bucket implementation for benchmark