--- a/search/materialize_test.go
+++ b/search/materialize_test.go
@@ -26,11 +26,11 @@ import (
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/prometheus/prometheus/util/teststorage"
 	"github.com/stretchr/testify/require"
-	"github.com/thanos-io/objstore/providers/filesystem"
 
 	"github.com/prometheus-community/parquet-common/convert"
 	"github.com/prometheus-community/parquet-common/schema"
 	"github.com/prometheus-community/parquet-common/storage"
+	"github.com/prometheus-community/parquet-common/util/filesystem"
 )
 
 func TestMaterializeE2E(t *testing.T) {