Noticed that ~50% of the benchmark time was spent doing syscall.Open. The filesystem bucket implementation we're using comes from thanos/objstore repository so I decided to copy it over and cache the file descriptors so that we can run get_range calls in parallel without getting throttled by the syscalls.

**Profile data**
- Original profile of the benchmark: https://flamegraph.com/share/207b00b6-450d-11f0-b170-826a90ca3d79
- After modifications: https://flamegraph.com/share/1cf80ed4-453d-11f0-9c9f-da4acad57e16
- Comparison between the two: https://flamegraph.com/share/2ef55dd9-453d-11f0-b170-826a90ca3d79/2f0948b9-453d-11f0-b170-826a90ca3d79

**Benchmark data**

<details>
<summary>Click to view benchstat results</summary>

```
❯ benchstat with-filesystem-bucket.out with-custom-filesystem-bucket-and-less-open-and-stat-calls.out
goos: darwin
goarch: arm64
pkg: github.com/prometheus-community/parquet-common/search
cpu: Apple M2 Pro
                                        │ with-filesystem-bucket.out │ with-custom-filesystem-bucket-and-less-open-and-stat-calls.out │
                                        │           sec/op           │                 sec/op                   vs base               │
Select/SingleMetricAllSeries-12                         530.7m ± ∞ ¹                              355.1m ± ∞ ¹  -33.09% (p=0.008 n=5)
Select/SingleMetricReducedSeries-12                     85.02m ± ∞ ¹                              24.58m ± ∞ ¹  -71.09% (p=0.008 n=5)
Select/SingleMetricOneSeries-12                         78.48m ± ∞ ¹                              22.44m ± ∞ ¹  -71.40% (p=0.008 n=5)
Select/SingleMetricSparseSeries-12                     132.96m ± ∞ ¹                              45.12m ± ∞ ¹  -66.06% (p=0.008 n=5)
Select/NonExistentSeries-12                             77.35m ± ∞ ¹                              21.75m ± ∞ ¹  -71.87% (p=0.008 n=5)
Select/MultipleMetricsRange-12                           1.860 ± ∞ ¹                               1.596 ± ∞ ¹  -14.17% (p=0.008 n=5)
Select/MultipleMetricsSparse-12                         669.2m ± ∞ ¹                              468.1m ± ∞ ¹  -30.05% (p=0.008 n=5)
Select/NegativeRegexSingleMetric-12                     411.8m ± ∞ ¹                              335.0m ± ∞ ¹  -18.65% (p=0.008 n=5)
Select/NegativeRegexMultipleMetrics-12                   1.314 ± ∞ ¹                               1.010 ± ∞ ¹  -23.13% (p=0.008 n=5)
Select/ExpensiveRegexSingleMetric-12                   155.81m ± ∞ ¹                              55.06m ± ∞ ¹  -64.66% (p=0.008 n=5)
Select/ExpensiveRegexMultipleMetrics-12                 488.3m ± ∞ ¹                              145.7m ± ∞ ¹  -70.17% (p=0.008 n=5)
geomean                                                 301.4m                                    139.5m        -53.71%
¹ need >= 6 samples for confidence interval at level 0.95

                                        │ with-filesystem-bucket.out │ with-custom-filesystem-bucket-and-less-open-and-stat-calls.out │
                                        │     bytes_get_range/op     │           bytes_get_range/op            vs base                │
Select/SingleMetricAllSeries-12                         9.137M ± ∞ ¹                             9.137M ± ∞ ¹  +0.00% (p=0.008 n=5)
Select/SingleMetricReducedSeries-12                     2.029M ± ∞ ¹                             2.029M ± ∞ ¹  +0.00% (p=0.008 n=5)
Select/SingleMetricOneSeries-12                         1.978M ± ∞ ¹                             1.978M ± ∞ ¹  -0.00% (p=0.008 n=5)
Select/SingleMetricSparseSeries-12                      5.487M ± ∞ ¹                             5.487M ± ∞ ¹  +0.00% (p=0.008 n=5)
Select/NonExistentSeries-12                             1.939M ± ∞ ¹                             1.939M ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/MultipleMetricsRange-12                          38.01M ± ∞ ¹                             38.01M ± ∞ ¹  +0.00% (p=0.008 n=5)
Select/MultipleMetricsSparse-12                         16.41M ± ∞ ¹                             16.41M ± ∞ ¹  +0.00% (p=0.008 n=5)
Select/NegativeRegexSingleMetric-12                     9.394M ± ∞ ¹                             9.394M ± ∞ ¹  +0.00% (p=0.008 n=5)
Select/NegativeRegexMultipleMetrics-12                  31.58M ± ∞ ¹                             31.58M ± ∞ ¹  -0.00% (p=0.008 n=5)
Select/ExpensiveRegexSingleMetric-12                    3.924M ± ∞ ¹                             3.924M ± ∞ ¹  -0.00% (p=0.008 n=5)
Select/ExpensiveRegexMultipleMetrics-12                 14.74M ± ∞ ¹                             14.74M ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                 7.492M                                   7.492M        +0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal

                                        │ with-filesystem-bucket.out │ with-custom-filesystem-bucket-and-less-open-and-stat-calls.out │
                                        │           get/op           │                 get/op                  vs base                │
Select/SingleMetricAllSeries-12                          0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricReducedSeries-12                      0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricOneSeries-12                          0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricSparseSeries-12                       0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NonExistentSeries-12                              0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/MultipleMetricsRange-12                           0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/MultipleMetricsSparse-12                          0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NegativeRegexSingleMetric-12                      0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NegativeRegexMultipleMetrics-12                   0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/ExpensiveRegexSingleMetric-12                     0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/ExpensiveRegexMultipleMetrics-12                  0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                            ³                                           +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                                        │ with-filesystem-bucket.out │ with-custom-filesystem-bucket-and-less-open-and-stat-calls.out │
                                        │        get_range/op        │              get_range/op               vs base                │
Select/SingleMetricAllSeries-12                         19.16k ± ∞ ¹                             19.16k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricReducedSeries-12                     10.93k ± ∞ ¹                             10.93k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricOneSeries-12                         10.90k ± ∞ ¹                             10.90k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricSparseSeries-12                      18.00k ± ∞ ¹                             18.00k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NonExistentSeries-12                             10.76k ± ∞ ¹                             10.76k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/MultipleMetricsRange-12                          72.41k ± ∞ ¹                             72.41k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/MultipleMetricsSparse-12                         47.21k ± ∞ ¹                             47.21k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NegativeRegexSingleMetric-12                     25.08k ± ∞ ¹                             25.08k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NegativeRegexMultipleMetrics-12                  81.76k ± ∞ ¹                             81.76k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/ExpensiveRegexSingleMetric-12                    18.69k ± ∞ ¹                             18.69k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/ExpensiveRegexMultipleMetrics-12                 62.10k ± ∞ ¹                             62.10k ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                 25.95k                                   25.95k        +0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal

                                        │ with-filesystem-bucket.out │ with-custom-filesystem-bucket-and-less-open-and-stat-calls.out │
                                        │         series/op          │               series/op                 vs base                │
Select/SingleMetricAllSeries-12                         300.0k ± ∞ ¹                             300.0k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricReducedSeries-12                     3.000k ± ∞ ¹                             3.000k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricOneSeries-12                          1.000 ± ∞ ¹                              1.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/SingleMetricSparseSeries-12                      5.000k ± ∞ ¹                             5.000k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NonExistentSeries-12                              0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/MultipleMetricsRange-12                          1.200M ± ∞ ¹                             1.200M ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/MultipleMetricsSparse-12                         300.0k ± ∞ ¹                             300.0k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NegativeRegexSingleMetric-12                     234.0k ± ∞ ¹                             234.0k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/NegativeRegexMultipleMetrics-12                  702.0k ± ∞ ¹                             702.0k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/ExpensiveRegexSingleMetric-12                    6.000k ± ∞ ¹                             6.000k ± ∞ ¹       ~ (p=1.000 n=5) ²
Select/ExpensiveRegexMultipleMetrics-12                  0.000 ± ∞ ¹                              0.000 ± ∞ ¹       ~ (p=1.000 n=5) ²
geomean                                                            ³                                           +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                                        │ with-filesystem-bucket.out │ with-custom-filesystem-bucket-and-less-open-and-stat-calls.out │
                                        │            B/op            │                   B/op                    vs base              │
Select/SingleMetricAllSeries-12                        787.3Mi ± ∞ ¹                              777.0Mi ± ∞ ¹  -1.30% (p=0.008 n=5)
Select/SingleMetricReducedSeries-12                    77.49Mi ± ∞ ¹                              71.65Mi ± ∞ ¹  -7.54% (p=0.008 n=5)
Select/SingleMetricOneSeries-12                        72.79Mi ± ∞ ¹                              66.96Mi ± ∞ ¹  -8.01% (p=0.008 n=5)
Select/SingleMetricSparseSeries-12                     140.8Mi ± ∞ ¹                              131.1Mi ± ∞ ¹  -6.83% (p=0.008 n=5)
Select/NonExistentSeries-12                            69.80Mi ± ∞ ¹                              64.05Mi ± ∞ ¹  -8.24% (p=0.008 n=5)
Select/MultipleMetricsRange-12                         3.146Gi ± ∞ ¹                              3.109Gi ± ∞ ¹  -1.20% (p=0.008 n=5)
Select/MultipleMetricsSparse-12                        1.058Gi ± ∞ ¹                              1.033Gi ± ∞ ¹  -2.33% (p=0.008 n=5)
Select/NegativeRegexSingleMetric-12                    729.6Mi ± ∞ ¹                              716.2Mi ± ∞ ¹  -1.84% (p=0.008 n=5)
Select/NegativeRegexMultipleMetrics-12                 2.213Gi ± ∞ ¹                              2.170Gi ± ∞ ¹  -1.93% (p=0.008 n=5)
Select/ExpensiveRegexSingleMetric-12                   202.7Mi ± ∞ ¹                              192.6Mi ± ∞ ¹  -4.94% (p=0.008 n=5)
Select/ExpensiveRegexMultipleMetrics-12                573.1Mi ± ∞ ¹                              539.9Mi ± ∞ ¹  -5.79% (p=0.008 n=5)
geomean                                                387.1Mi                                    369.3Mi        -4.58%
¹ need >= 6 samples for confidence interval at level 0.95

                                        │ with-filesystem-bucket.out │ with-custom-filesystem-bucket-and-less-open-and-stat-calls.out │
                                        │         allocs/op          │                allocs/op                 vs base               │
Select/SingleMetricAllSeries-12                         9.960M ± ∞ ¹                              9.845M ± ∞ ¹   -1.15% (p=0.008 n=5)
Select/SingleMetricReducedSeries-12                     355.1k ± ∞ ¹                              289.6k ± ∞ ¹  -18.46% (p=0.008 n=5)
Select/SingleMetricOneSeries-12                         331.7k ± ∞ ¹                              266.2k ± ∞ ¹  -19.73% (p=0.008 n=5)
Select/SingleMetricSparseSeries-12                      755.9k ± ∞ ¹                              647.9k ± ∞ ¹  -14.29% (p=0.008 n=5)
Select/NonExistentSeries-12                             256.1k ± ∞ ¹                              191.6k ± ∞ ¹  -25.21% (p=0.008 n=5)
Select/MultipleMetricsRange-12                          40.14M ± ∞ ¹                              39.71M ± ∞ ¹   -1.08% (p=0.008 n=5)
Select/MultipleMetricsSparse-12                         11.10M ± ∞ ¹                              10.82M ± ∞ ¹   -2.55% (p=0.008 n=5)
Select/NegativeRegexSingleMetric-12                     8.504M ± ∞ ¹                              8.353M ± ∞ ¹   -1.77% (p=0.008 n=5)
Select/NegativeRegexMultipleMetrics-12                  24.99M ± ∞ ¹                              24.50M ± ∞ ¹   -1.96% (p=0.008 n=5)
Select/ExpensiveRegexSingleMetric-12                   1027.8k ± ∞ ¹                              915.7k ± ∞ ¹  -10.92% (p=0.008 n=5)
Select/ExpensiveRegexMultipleMetrics-12                 2.168M ± ∞ ¹                              1.796M ± ∞ ¹  -17.19% (p=0.008 n=5)
geomean                                                 2.664M                                    2.376M        -10.81%
¹ need >= 6 samples for confidence interval at level 0.95
```
</details>

Summary:
* Geomean shows 53.71% improvement so about 2x faster.
* 10.81% fewer allocations and 4.6% less memory usage.
* The benchmark cases that benefited the more were:
  * SingleMetricReducedSeries: -71.09% (85ms → 24.6ms)
  * SingleMetricOneSeries: -71.40% (78.5ms → 22.4ms)
  * NonExistentSeries: -71.87% (77.4ms → 21.8ms)
  * ExpensiveRegexMultipleMetrics: -70.17% (488ms → 146ms)

Note: I also tried the InMemBucket implementation but it did deadlock and I figured maybe filesystem with safe ReadAt would be easier to achieve these performance improvements. Also closer to what it would look like in our systems I think.