@npazosmendez 2025-06-09
We could also tackle this by benchmarking using the on-disk `ParquetFile`: https://github.com/prometheus-community/parquet-common/pull/58 . It adds the `io.Closer` interface to the ReadAt used by the ParquetFiles, to manage the lifetime of the file. It should be as simple as changing the Shard constructor

(Note that I might close that one in favor of https://github.com/prometheus-community/parquet-common/pull/64, which works it further)